Commenting, changed name of change_stance
Changed the function change_stance of door to be alled open_or_close.

Commented on some things to make them clearer.

Added a drawDoor function to the visualization that is called if there
is a door. Hopefully there are no errors cause I did not have time to
compile this code on the server.