Worked on copy_state method
The copy_state method now copies all of the dictionary elements from the
state to the new State.

The copy state then does some checks to place in the correct doors into
the walls.

Made the ROOMS and DOORS constants not refer to the initial state. I
think the variable name should be changed then.