Bug fixing, infastructure of Puzzle Role
There is now only a drawPuzzle function called in the visualization.
This is the only place where the puzzle is constructed visually.

The drawPuzzle function calls "render()" automatically so that no longer
needs to be included in evaluation strings.

The drawPuzzle function also deals with puzzle the correct url on the
canvas. However more work needs to be done on it to get it to correctly
backtrack to a state that did not have a Puzzle.

Fixed a couple bugs here and there.

Removed unnecessary comments.

Image now works (kind of) with backtracking.