Work on getting Music to work
Starting to port things from test files to actual files. The music stuff
will be displayed with a div so i can have a playMusic button. I changed
the display types of the canvas, svg, and musicDisplay to be block
because the musicDisplay div wouldn't show up on display "initial" and I
wanted consistency