Made Room operator use callbacks
I will soon change all operators to use callbacks that give the client a
new state to use. This will move more work into the vis file and will
also allow more asynchronous actions to do be with operators. For
example the creation of a menu where the user selects objects.  This is
a great example of how design DOES often inform structure. The whole
foundation of operators had to be re-written for a DESIGN choice to help
the user!!